initial commit for identity addition
right now this is using local keys that I generated.  to use it you will
need my private key.  we need to get garrett's key from google working
so we can all work on it.  Also, there is a bug in this.  I have to
click the extension twice for it to update the json blob with email and
user id