Added a flag to set whether scraping the users history occurs every time or only for new users.
To centralize the initialization of firebase, ensure the initialization always occurs before use and reduce the amount of global scope we use, the firebase api and database are now wrapped in promises.

Fixed up some of the source entries and added several.

Created a function for handling changing the read count.

Added an Article class.

Article records are now only written if no record already exists or the new record is a full record.

The history search now retrieves a maximum of 10000 items.

background.js is now responsible for determining when to write to the database; the "database_write" message type has been replaced with "update_current_article" which allows the content script to push scraped information into the background.

The background now listens for when tabs update or close to decide to write article data. We are now tracking the amount of content the user scrolls through and it is highly infeasible to update the database on each of these events, which may occur up to 100 times per page, so it was necessary to select a cut-off for when it was appropriate to write to the database- navigating away from the page seemed a reasonable choice as it indicates the user no longer wishes to interact with that page for now.