Removed the addition of the hashcode to String, it is no longer used.
Removed the sources object, now the sources are retrieved via a message sent to the background page. This creates a single, authoritative location for all source declarations and avoids the hassle and danger of maintaining them in multiple files.

The article data is now pushed to the backend via sending a runtime message.