Popup.js no longer accesses the database directly to get article information. Now, to ensure that there is a single "canonical" set of article data throughout the extension, the article data for the current page is accessible in the background page wrapped in a promise called current_article, can be retrieved by content scripts by sending a message with the "getCurrentArticle" type and can be updated by passing a message with the "updateCurrentArticle" type and passing the entire new article information as the message.
"Article data" represents the data about the article itself, such as text, author, etc. including the ids of all users who have read it. The user metadata is information related to the article that is particular to a certain user, including their lean and star ratings and the last date they read the article.